Handler/Serializer root classes (no IEntity Dependency)
Non-breaking change, adding a root class , so you can bypass the dependency on IEntity.
